% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourplots.R
\name{fourplots}
\alias{fourplots}
\title{fourplots}
\usage{
fourplots(x)
}
\arguments{
\item{x}{the datavector}
}
\value{
Returns the four plots in one object
}
\description{
Given a dataset it makes 4 different plot
}
\details{
It the first plot is the data, the second plot is the differences, the third plot a histogram and the fourth a qq-plot
}
\examples{
fourplots(1:10)

}
